WIP: Move ApplicantQuestion inner classes to upper classes
