+++ universal-application-tool-0.0.1/app/services/applicant/question/TextQuestion.java
@@ -1,0 +1,91 @@
+package services.applicant.question;
+
+import com.google.common.collect.ImmutableSet;
+import services.Path;
+import services.applicant.ValidationErrorMessage;
+import services.question.QuestionType;
+import services.question.TextQuestionDefinition;
+
+import java.util.Optional;
+
+public class TextQuestion implements PresentsErrors {
+
+  private final ApplicantQuestion applicantQuestion;
+  private Optional<String> textValue;
+
+  public TextQuestion(ApplicantQuestion applicantQuestion) {
+    this.applicantQuestion = applicantQuestion;
+    assertQuestionType();
+  }
+
+  @Override
+  public boolean hasQuestionErrors() {
+    return !getQuestionErrors().isEmpty();
+  }
+
+  public ImmutableSet<ValidationErrorMessage> getQuestionErrors() {
+    // TODO(https://github.com/seattle-uat/civiform/issues/634): Fix bug related to hasValue.
+    if (!hasValue()) {
+      return ImmutableSet.of();
+    }
+
+    TextQuestionDefinition definition = getQuestionDefinition();
+    int textLength = getTextValue().map(s -> s.length()).orElse(0);
+    ImmutableSet.Builder<ValidationErrorMessage> errors =
+            ImmutableSet.<ValidationErrorMessage>builder();
+
+    if (definition.getMinLength().isPresent()) {
+      int minLength = definition.getMinLength().getAsInt();
+      if (textLength < minLength) {
+        errors.add(ValidationErrorMessage.textTooShortError(minLength));
+      }
+    }
+
+    if (definition.getMaxLength().isPresent()) {
+      int maxLength = definition.getMaxLength().getAsInt();
+      if (textLength > maxLength) {
+        errors.add(ValidationErrorMessage.textTooLongError(maxLength));
+      }
+    }
+
+    return errors.build();
+  }
+
+  @Override
+  public boolean hasTypeSpecificErrors() {
+    // There are no inherent requirements in a text question.
+    return false;
+  }
+
+  public boolean hasValue() {
+    return getTextValue().isPresent();
+  }
+
+  public Optional<String> getTextValue() {
+    if (textValue != null) {
+      return textValue;
+    }
+
+    textValue = applicantQuestion.applicantData.readString(getTextPath());
+
+    return textValue;
+  }
+
+  public void assertQuestionType() {
+    if (!applicantQuestion.getType().equals(QuestionType.TEXT)) {
+      throw new RuntimeException(
+              String.format(
+                      "Question is not a TEXT question: %s (type: %s)",
+                      applicantQuestion.questionDefinition.getPath(), applicantQuestion.questionDefinition.getQuestionType()));
+    }
+  }
+
+  public TextQuestionDefinition getQuestionDefinition() {
+    assertQuestionType();
+    return (TextQuestionDefinition) applicantQuestion.questionDefinition;
+  }
+
+  public Path getTextPath() {
+    return getQuestionDefinition().getTextPath();
+  }
+}
