+++ universal-application-tool-0.0.1/app/views/applicant/ApplicantProgramBlockEditView.java
@@ -14,7 +14,7 @@ import play.i18n.Messages;
 import play.mvc.Http;
 import play.mvc.Http.HttpVerbs;
 import play.twirl.api.Content;
-import services.applicant.ApplicantQuestion;
+import services.applicant.question.ApplicantQuestion;
 import services.applicant.Block;
 import views.BaseHtmlView;
 import views.questiontypes.ApplicantQuestionRendererFactory;
