+++ universal-application-tool-0.0.1/app/views/questiontypes/AddressQuestionRenderer.java
@@ -4,7 +4,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static j2html.TagCreator.div;
 
 import j2html.tags.Tag;
-import services.applicant.ApplicantQuestion;
+import services.applicant.question.ApplicantQuestion;
+import services.applicant.question.AddressQuestion;
 import views.BaseHtmlView;
 import views.components.FieldWithLabel;
 import views.style.ReferenceClasses;
@@ -24,7 +25,7 @@ public class AddressQuestionRenderer extends BaseHtmlView implements ApplicantQu
 
   @Override
   public Tag render() {
-    ApplicantQuestion.AddressQuestion addressQuestion = question.getAddressQuestion();
+    AddressQuestion addressQuestion = question.getAddressQuestion();
 
     return div()
         .withId(question.getPath().path())
