+++ universal-application-tool-0.0.1/app/views/questiontypes/NameQuestionRenderer.java
@@ -6,7 +6,7 @@ import static j2html.TagCreator.input;
 import static j2html.TagCreator.label;
 
 import j2html.tags.Tag;
-import services.applicant.ApplicantQuestion;
+import services.applicant.question.ApplicantQuestion;
 import views.BaseHtmlView;
 import views.style.ReferenceClasses;
 import views.style.Styles;
