+++ universal-application-tool-0.0.1/app/views/questiontypes/NumberQuestionRenderer.java
@@ -5,7 +5,8 @@ import static j2html.TagCreator.div;
 import static j2html.TagCreator.input;
 
 import j2html.tags.Tag;
-import services.applicant.ApplicantQuestion;
+import services.applicant.question.ApplicantQuestion;
+import services.applicant.question.NumberQuestion;
 import views.BaseHtmlView;
 import views.style.ReferenceClasses;
 import views.style.Styles;
@@ -20,7 +21,7 @@ public class NumberQuestionRenderer extends BaseHtmlView implements ApplicantQue
 
   @Override
   public Tag render() {
-    ApplicantQuestion.NumberQuestion numberQuestion = question.getNumberQuestion();
+    NumberQuestion numberQuestion = question.getNumberQuestion();
 
     return div()
         .withId(question.getPath().path())
