+++ universal-application-tool-0.0.1/test/views/questiontypes/TextQuestionRendererTest.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 import repository.WithPostgresContainer;
 import services.Path;
 import services.applicant.ApplicantData;
-import services.applicant.ApplicantQuestion;
+import services.applicant.question.ApplicantQuestion;
 import services.question.TextQuestionDefinition;
 import services.question.TextQuestionDefinition.TextValidationPredicates;
